refactor(userspace-dp): #6386 extract prerouting_scope sibling
Move the #5802 pre-routing NAT ingress-scope resolver out of
poll_descriptor/mod.rs into a new pub(super) sibling module
prerouting_scope.rs: struct PreroutingIngressScope and
prerouting_ingress_scope, plus the prerouting_scope_tests suite
(loaded from the sibling via #[path], the reject_reply.rs convention).

Pure code motion. Both items were already pub(super); the struct/fn
bodies are byte-identical to their prior location (verified with
git diff --color-moved). The only non-motion change is the added
#[inline] on prerouting_ingress_scope, restoring same-CGU inlining
eligibility across the new module boundary per the #6386 hot-path
preservation contract. mod.rs re-imports both names via
`use prerouting_scope::{PreroutingIngressScope, prerouting_ingress_scope};`
so every loop call site is textually unchanged.

The one required non-verbatim test change: prerouting_scope_tests's
`super::super::test_fixtures::nat_snapshot()` becomes the absolute
`crate::afxdp::test_fixtures::nat_snapshot()` — the new module depth
makes the old relative path resolve to poll_descriptor rather than
afxdp.

Validation: cargo build (release) clean; cargo test --no-run clean;
moved bodies byte-identical modulo the #[inline] and the nat_snapshot
path fix.
